refactor: split files in folders
