--- src/components/Page404.tsx
@@ -1,3 +1,0 @@
-export default function UncontrolledForm() {
-  return <h1>404</h1>;
-}
