+++ src/pages/Page404.tsx
@@ -1,0 +1,3 @@
+export default function Page404() {
+  return <h1>404</h1>;
+}
